--- Payrails/Classes/Public/Views/Card/PaymentButton.swift
@@ -1,5 +1,0 @@
-import UIKit
-
-class PaymentButton: ActionButton {
-
-}
